pyproject.toml
[tool.poetry]
name = "snake-game"
version = "0.1.0"
description = ""
authors = ["Erick Duarte <erickod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.7.4"
python-socketio = {version = "4.6.0", extras = ["asyncio_client"]}
python-engineio = "3.13.2"
pyngrok = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
